+++ src/helper/TextDocumentHighlighter.ts
@@ -2,14 +2,14 @@ import vscode, { TextEditor, TextEditorDecorationType } from 'vscode'
 import { SensorValues, UnifiedPath } from '@oaklean/profiler-core'
 
 import { calcOrReturnSensorValue } from './FormulaHelper'
-import { NumberHelper } from './NumberHelper'
+import { SensorValueFormatHelper } from './SensorValueFormatHelper'
 
 import { pad } from '../system/string'
 import { getImportanceColor } from '../system/color'
 import { GlyphChars } from '../constants/characters'
 import { PROFILE_PERCENT_PRECISION } from '../constants/profile'
 import { Container } from '../container'
-import { ExtendedSensorValueType, SensorValueTypeNames, UnitPerSensorValue } from '../types/sensorValues'
+import { ExtendedSensorValueType, SensorValueTypeNames } from '../types/sensorValues'
 import { Profile } from '../types/profile'
 import { Color } from '../types/color'
 import { SensorValueRepresentation } from '../types/sensorValueRepresentation'
@@ -102,21 +102,25 @@ export class TextDocumentHighlighter {
 				const formula = sensorValueRepresentation.formula
 				const calculatedFormula = calcOrReturnSensorValue(
 					sourceNodeMetaData.sensorValues, selectedSensorValueType, formula)
-				const roundedCalculatedFormula = NumberHelper.round(calculatedFormula, selectedSensorValueType, 
-					UnitPerSensorValue[selectedSensorValueType])
+				const formattedCalculatedFormula = SensorValueFormatHelper.format(
+					calculatedFormula,
+					selectedSensorValueType
+				)
 				const formulaTotal = calcOrReturnSensorValue(
 					totalAndMaxMetaData.total.sensorValues, selectedSensorValueType, formula)
 				const relativeToToalForFormula = calculatedFormula / formulaTotal
-				message = `${formula}: ${roundedCalculatedFormula.newValue} ` +
+				message = `${formula}: ${formattedCalculatedFormula.value} ` +
 					`(${relativeToToalForFormula.toFixed(PROFILE_PERCENT_PRECISION)}%)`
 				weigth = calculatedFormula / calcOrReturnSensorValue(
 					totalAndMaxMetaData.max.sensorValues, selectedSensorValueType, formula)
 			} else {
-				const roundedValue = NumberHelper.round(value, selectedSensorValueType, 
-					UnitPerSensorValue[selectedSensorValueType])
+				const formattedValue = SensorValueFormatHelper.format(
+					value,
+					selectedSensorValueType
+				)
 				message =
 					SensorValueTypeNames[selectedSensorValueType] +
-					`: ${roundedValue.newValue} ${roundedValue.newUnit} ` +
+					`: ${formattedValue.value} ${formattedValue.unit} ` +
 					`(${relativeToToal.toFixed(PROFILE_PERCENT_PRECISION)}%)`
 				weigth = value / totalAndMaxMetaData.max.sensorValues[selectedSensorValueType]
 			}